File partitions

+++ Chaining/Headerchain/ChainInserter.cs
@@ -89,7 +89,8 @@ namespace BToken.Chaining
       void FindPreviousHeader(NetworkHeader header)
       {
         Probe.Chain = Headerchain.MainChain;
-        if (Probe.GoTo(header.HashPrevious, Headerchain.MainChain.HeaderRoot)) { return; }
+        if (Probe.GoTo(header.HashPrevious, Headerchain.MainChain.HeaderRoot))
+        { return; }
 
         foreach (Chain chain in Headerchain.SecondaryChains)
         {
@@ -108,9 +109,9 @@ namespace BToken.Chaining
       }
       void ValidateCheckpoint(UInt256 headerHash)
       {
-        uint nextHeaderHeight = Probe.GetHeight() + 1;
+        int nextHeaderHeight = Probe.GetHeight() + 1;
 
-        uint highestCheckpointHight = Headerchain.Checkpoints.Max(x => x.Height);
+        int highestCheckpointHight = Headerchain.Checkpoints.Max(x => x.Height);
         bool mainChainLongerThanHighestCheckpoint = highestCheckpointHight <= Headerchain.MainChain.Height;
         bool nextHeightBelowHighestCheckpoint = nextHeaderHeight <= highestCheckpointHight;
         if (mainChainLongerThanHighestCheckpoint && nextHeightBelowHighestCheckpoint)
@@ -123,7 +124,7 @@ namespace BToken.Chaining
           throw new ChainException(ChainCode.INVALID);
         }
       }
-      bool ValidateBlockLocation(uint height, UInt256 hash)
+      bool ValidateBlockLocation(int height, UInt256 hash)
       {
         HeaderLocation checkpoint = Headerchain.Checkpoints.Find(c => c.Height == height);
         if (checkpoint != null)
@@ -135,8 +136,8 @@ namespace BToken.Chaining
       }
       void ValidateProofOfWork(uint nBits, UInt256 headerHash)
       {
-        uint nextHeight = Probe.GetHeight() + 1;
-        if (nBits != TargetManager.GetNextTargetBits(Probe.Header, nextHeight))
+        int nextHeight = Probe.GetHeight() + 1;
+        if (nBits != TargetManager.GetNextTargetBits(Probe.Header, (uint)nextHeight))
         {
           throw new ChainException(ChainCode.INVALID);
         }
